refactor(cli): rename schema-diff to breaking-changes
Address maintainer feedback that schema-diff implies a generic diff
rather than breaking-change detection. Rename the CLI command, module,
and public utils path to breaking_changes / breaking-changes.
