+++ RELEASE.md
@@ -10,5 +10,5 @@ social_messages:
 
 This release adds a utility and CLI for detecting breaking GraphQL schema changes between SDL documents.
 
-- `strawberry.utils.schema_diff.find_breaking_changes_between_sdls` wraps graphql-core's `find_breaking_changes` for two SDL strings
-- `strawberry schema-diff` compares two `.graphql` files and exits 0 (no breaking changes), 1 (breaking changes found), or 2 (parse/read errors)
+- `strawberry.utils.breaking_changes.find_breaking_changes_between_sdls` wraps graphql-core's `find_breaking_changes` for two SDL strings
+- `strawberry breaking-changes` compares two `.graphql` files and exits 0 (no breaking changes), 1 (breaking changes found), or 2 (parse/read errors)
